src/requirements.txt
tensorflow
keras
medicai
pillow
numpy
matplotlib
pandas
